NeuraTuneBE/requirements.txt
billboard.py==7.0.1
Django==4.2.2
librosa==0.10.1
lyricsgenius==3.0.1
numpy==1.24.3
pandas==2.2.2
tensorflow==2.13.0
tensorflow_intel==2.13.0
transformers==4.41.2
tensorflow_intel==2.14.0
tensorflow_intel==2.16.1
